requirements.txt
Flask==3.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
reportlab==4.0.7
Pillow>=10.0.0
python-barcode==0.15.1
openpyxl==3.1.2
gunicorn==21.2.0

